Renamed logpicker to logfetcher.
