+++ pkg/mapserver/logfetcher/logfetcher_test.go
@@ -1,4 +1,4 @@
-package logpicker
+package logfetcher
 
 import (
 	"context"
